fix(v1): resolve relative links in documents properly
Updated the "mdToHtml" function so that it produces a map from all documents.
It was set to create map only for "next release" docs by checking if the docs
metadata contained "original_id" or not. It's now updated to create map for
versioned documents also. In addition, updated the key of the map to reflect
actual directory pattern. This allows picking up the right key in "mdToHtmlify"
function in "docs.js" by resolving the relative URL.